move discernValBool
