+++ pkg/osquery/tables/firmwarepasswd/parser.go
@@ -3,11 +3,9 @@ package firmwarepasswd
 import (
 	"bufio"
 	"bytes"
-	"strings"
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
-	"github.com/pkg/errors"
 )
 
 type Matcher struct {
@@ -81,15 +79,4 @@ func (p *OutputParser) Parse(input *bytes.Buffer) []map[string]string {
 		level.Debug(p.logger).Log("msg", "scanner error", "err", err)
 	}
 	return results
-}
-
-func discernValBool(in string) (bool, error) {
-	switch strings.TrimSpace(strings.ToLower(in)) {
-	case "true", "t", "1", "y", "yes":
-		return true, nil
-	case "false", "f", "0", "n", "no":
-		return false, nil
-	}
-
-	return false, errors.Errorf("Can't discern boolean from string <%s>", in)
 }
