+++ pkg/osquery/tables/firmwarepasswd/testdata/check-garbage.txt
@@ -1,0 +1,3 @@
+Hello: world
+Password Enabled: No
+Unknown: xxx
